+++ src/headless/plugins/muc/index.js
@@ -236,6 +236,7 @@ converse.plugins.add('converse-muc', {
             'auto_register_muc_nickname': false,
             'hide_muc_participants': false,
             'locked_muc_domain': false,
+            'modtools_disable_assign': false,
             'muc_clear_messages_on_leave': true,
             'muc_domain': undefined,
             'muc_fetch_members': true,
