+++ src/plugins/chatview/heading.js
@@ -4,7 +4,7 @@ import { CustomElement } from 'shared/components/element.js';
 import { __ } from 'i18n';
 import { _converse, api } from "@converse/headless/core";
 
-import './styles//chat-head.scss';
+import './styles/chat-head.scss';
 
 
 export default class ChatHeading extends CustomElement {
