+++ src/plugins/muc-views/templates/moderator-tools.js
@@ -1,7 +1,6 @@
 import spinner from "templates/spinner.js";
 import { __ } from 'i18n';
 import { html } from "lit";
-import { modal_header_close_button } from "modals/templates/buttons.js"
 
 
 function getRoleHelpText (role) {
@@ -135,13 +134,13 @@ const affiliation_list_item = (o) => html`
 `;
 
 
-const tpl_navigation = (o) => html`
+const tpl_navigation = () => html`
     <ul class="nav nav-pills justify-content-center">
         <li role="presentation" class="nav-item">
-            <a class="nav-link active" id="affiliations-tab" href="#affiliations-tabpanel" aria-controls="affiliations-tabpanel" role="tab" data-toggle="tab" @click=${o.switchTab}>Affiliations</a>
+            <a class="nav-link active" id="affiliations-tab" href="#affiliations-tabpanel" aria-controls="affiliations-tabpanel" role="tab" data-toggle="tab">Affiliations</a>
         </li>
         <li role="presentation" class="nav-item">
-            <a class="nav-link" id="roles-tab" href="#roles-tabpanel" aria-controls="roles-tabpanel" role="tab" data-toggle="tab" @click=${o.switchTab}>Roles</a>
+            <a class="nav-link" id="roles-tab" href="#roles-tabpanel" aria-controls="roles-tabpanel" role="tab" data-toggle="tab">Roles</a>
         </li>
     </ul>
 `;
@@ -149,7 +148,6 @@ const tpl_navigation = (o) => html`
 
 export default (o) => {
     const i18n_affiliation = __('Affiliation');
-    const i18n_moderator_tools = __('Moderator Tools');
     const i18n_no_users_with_aff = __('No users with that affiliation found.')
     const i18n_no_users_with_role = __('No users with that role found.');
     const i18n_filter = __('Type here to filter the search results');
@@ -167,94 +165,83 @@ export default (o) => {
     );
     const show_both_tabs = o.queryable_roles.length && o.queryable_affiliations.length;
     return html`
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="converse-modtools-modal-label">${i18n_moderator_tools}</h5>
-                ${modal_header_close_button}
-            </div>
-            <div class="modal-body d-flex flex-column">
-                <span class="modal-alert"></span>
+        ${o.alert_message ? html`<div class="alert alert-${o.alert_type}" role="alert">${o.alert_message}</div>` : '' }
+        ${ show_both_tabs ? tpl_navigation() : '' }
 
-                ${ show_both_tabs ? tpl_navigation(o) : '' }
-
-                <div class="tab-content">
-                    <div class="tab-pane tab-pane--columns ${ o.queryable_affiliations.length ? 'active' : ''}" id="affiliations-tabpanel" role="tabpanel" aria-labelledby="affiliations-tab">
-                        <form class="converse-form query-affiliation" @submit=${o.queryAffiliation}>
-                            <p class="helptext pb-3">${i18n_helptext_affiliation}</p>
-                            <div class="form-group">
-                                <label for="affiliation">
-                                    <strong>${i18n_affiliation}:</strong>
-                                </label>
-                                <div class="row">
-                                    <div class="col">
-                                        <select class="custom-select select-affiliation" name="affiliation">
-                                            ${o.queryable_affiliations.map(item => affiliation_option(Object.assign({item}, o)))}
-                                        </select>
-                                    </div>
-                                    <div class="col">
-                                        <input type="submit" class="btn btn-primary" name="users_with_affiliation" value="${i18n_show_users}"/>
-                                    </div>
-                                </div>
-                                <div class="row">
-                                    <div class="col mt-3">
-                                        ${ (Array.isArray(o.users_with_affiliation) && o.users_with_affiliation.length > 5) ?
-                                            html`<input class="form-control" .value="${o.affiliations_filter}" @keyup=${o.filterAffiliationResults} type="text" name="filter" placeholder="${i18n_filter}"/>` : '' }
-                                    </div>
-                                </div>
-
-                                ${ getAffiliationHelpText(o.affiliation) ?
-                                    html`<div class="row"><div class="col pt-2"><p class="helptext pb-3">${getAffiliationHelpText(o.affiliation)}</p></div></div>` : '' }
+        <div class="tab-content">
+            <div class="tab-pane tab-pane--columns ${ o.queryable_affiliations.length ? 'active' : ''}" id="affiliations-tabpanel" role="tabpanel" aria-labelledby="affiliations-tab">
+                <form class="converse-form query-affiliation" @submit=${o.queryAffiliation}>
+                    <p class="helptext pb-3">${i18n_helptext_affiliation}</p>
+                    <div class="form-group">
+                        <label for="affiliation">
+                            <strong>${i18n_affiliation}:</strong>
+                        </label>
+                        <div class="row">
+                            <div class="col">
+                                <select class="custom-select select-affiliation" name="affiliation">
+                                    ${o.queryable_affiliations.map(item => affiliation_option(Object.assign({item}, o)))}
+                                </select>
                             </div>
-                        </form>
-                        <div class="scrollable-container">
-                            <ul class="list-group list-group--users">
-                                ${ (o.loading_users_with_affiliation) ? html`<li class="list-group-item"> ${spinner()} </li>` : '' }
-                                ${ (Array.isArray(o.users_with_affiliation) && o.users_with_affiliation.length === 0) ?
-                                        html`<li class="list-group-item">${i18n_no_users_with_aff}</li>` : '' }
-
-                                ${ (o.users_with_affiliation instanceof Error) ?
-                                        html`<li class="list-group-item">${o.users_with_affiliation.message}</li>` :
-                                        (o.users_with_affiliation || []).map(item => ((item.nick || item.jid).match(new RegExp(o.affiliations_filter, 'i')) ? affiliation_list_item(Object.assign({item}, o)) : '')) }
-                            </ul>
-                        </div>
-                    </div>
-
-                    <div class="tab-pane tab-pane--columns ${ !show_both_tabs && o.queryable_roles.length ? 'active' : ''}" id="roles-tabpanel" role="tabpanel" aria-labelledby="roles-tab">
-                        <form class="converse-form query-role" @submit=${o.queryRole}>
-                            <p class="helptext pb-3">${i18n_helptext_role}</p>
-                            <div class="form-group">
-                                <label for="role"><strong>${i18n_role}:</strong></label>
-                                <div class="row">
-                                    <div class="col">
-                                        <select class="custom-select select-role" name="role">
-                                            ${o.queryable_roles.map(item => role_option(Object.assign({item}, o)))}
-                                        </select>
-                                    </div>
-                                    <div class="col">
-                                        <input type="submit" class="btn btn-primary" name="users_with_role" value="${i18n_show_users}"/>
-                                    </div>
-                                </div>
-                                <div class="row">
-                                    <div class="col mt-3">
-                                        ${ (Array.isArray(o.users_with_role) && o.users_with_role.length > 5) ?
-                                            html`<input class="form-control" .value="${o.roles_filter}" @keyup=${o.filterRoleResults} type="text" name="filter" placeholder="${i18n_filter}"/>` : '' }
-                                    </div>
-                                </div>
-
-                                ${ getRoleHelpText(o.role) ? html`<div class="row"><div class="col pt-2"><p class="helptext pb-3">${getRoleHelpText(o.role)}</p></div></div>` :  ''}
+                            <div class="col">
+                                <input type="submit" class="btn btn-primary" name="users_with_affiliation" value="${i18n_show_users}"/>
                             </div>
-                        </form>
-                        <div class="scrollable-container">
-                            <ul class="list-group list-group--users">
-                                ${ o.loading_users_with_role ? html`<li class="list-group-item"> ${spinner()} </li>` : '' }
-                                ${ (o.users_with_role && o.users_with_role.length === 0) ? html`<li class="list-group-item">${i18n_no_users_with_role}</li>` : '' }
-                                ${ (o.users_with_role || []).map(item => (item.nick.match(o.roles_filter) ? role_list_item(Object.assign({item}, o)) : '')) }
-                            </ul>
                         </div>
+                        <div class="row">
+                            <div class="col mt-3">
+                                ${ (Array.isArray(o.users_with_affiliation) && o.users_with_affiliation.length > 5) ?
+                                    html`<input class="form-control" .value="${o.affiliations_filter}" @keyup=${o.filterAffiliationResults} type="text" name="filter" placeholder="${i18n_filter}"/>` : '' }
+                            </div>
+                        </div>
+
+                        ${ getAffiliationHelpText(o.affiliation) ?
+                            html`<div class="row"><div class="col pt-2"><p class="helptext pb-3">${getAffiliationHelpText(o.affiliation)}</p></div></div>` : '' }
                     </div>
+                </form>
+                <div class="scrollable-container">
+                    <ul class="list-group list-group--users">
+                        ${ (o.loading_users_with_affiliation) ? html`<li class="list-group-item"> ${spinner()} </li>` : '' }
+                        ${ (Array.isArray(o.users_with_affiliation) && o.users_with_affiliation.length === 0) ?
+                                html`<li class="list-group-item">${i18n_no_users_with_aff}</li>` : '' }
+
+                        ${ (o.users_with_affiliation instanceof Error) ?
+                                html`<li class="list-group-item">${o.users_with_affiliation.message}</li>` :
+                                (o.users_with_affiliation || []).map(item => ((item.nick || item.jid).match(new RegExp(o.affiliations_filter, 'i')) ? affiliation_list_item(Object.assign({item}, o)) : '')) }
+                    </ul>
                 </div>
             </div>
-        </div>
-    </div>`;
+
+            <div class="tab-pane tab-pane--columns ${ !show_both_tabs && o.queryable_roles.length ? 'active' : ''}" id="roles-tabpanel" role="tabpanel" aria-labelledby="roles-tab">
+                <form class="converse-form query-role" @submit=${o.queryRole}>
+                    <p class="helptext pb-3">${i18n_helptext_role}</p>
+                    <div class="form-group">
+                        <label for="role"><strong>${i18n_role}:</strong></label>
+                        <div class="row">
+                            <div class="col">
+                                <select class="custom-select select-role" name="role">
+                                    ${o.queryable_roles.map(item => role_option(Object.assign({item}, o)))}
+                                </select>
+                            </div>
+                            <div class="col">
+                                <input type="submit" class="btn btn-primary" name="users_with_role" value="${i18n_show_users}"/>
+                            </div>
+                        </div>
+                        <div class="row">
+                            <div class="col mt-3">
+                                ${ (Array.isArray(o.users_with_role) && o.users_with_role.length > 5) ?
+                                    html`<input class="form-control" .value="${o.roles_filter}" @keyup=${o.filterRoleResults} type="text" name="filter" placeholder="${i18n_filter}"/>` : '' }
+                            </div>
+                        </div>
+
+                        ${ getRoleHelpText(o.role) ? html`<div class="row"><div class="col pt-2"><p class="helptext pb-3">${getRoleHelpText(o.role)}</p></div></div>` :  ''}
+                    </div>
+                </form>
+                <div class="scrollable-container">
+                    <ul class="list-group list-group--users">
+                        ${ o.loading_users_with_role ? html`<li class="list-group-item"> ${spinner()} </li>` : '' }
+                        ${ (o.users_with_role && o.users_with_role.length === 0) ? html`<li class="list-group-item">${i18n_no_users_with_role}</li>` : '' }
+                        ${ (o.users_with_role || []).map(item => (item.nick.match(o.roles_filter) ? role_list_item(Object.assign({item}, o)) : '')) }
+                    </ul>
+                </div>
+            </div>
+        </div>`;
 }
