+++ webpack.html
@@ -27,8 +27,8 @@
             message_limit: 300,
             auto_register_muc_nickname: true,
             loglevel: 'debug',
-            modtools_disable_assign: ['owner', 'moderator', 'participant', 'visitor'],
-            modtools_disable_query: ['moderator', 'participant', 'visitor'],
+            // modtools_disable_assign: ['owner', 'moderator', 'participant', 'visitor'],
+            // modtools_disable_query: ['moderator', 'participant', 'visitor'],
             enable_smacks: true,
             // connection_options: { 'worker': '/dist/shared-connection-worker.js' },
             // persistent_store: 'IndexedDB',
